requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# CDDB/src/main.py: 28
python_dotenv == 0.15.0

# CDDB/src/daily.py: 8
# CDDB/src/main.py: 26
python_telegram_bot == 13.3
